+++ example/angular2-try/components/SsqBobao.js
@@ -11,7 +11,7 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 };
 var __param = (this && this.__param) || function (paramIndex, decorator) {
     return function (target, key) { decorator(target, key, paramIndex); }
-};System.register(['angular2/angular2', 'angular2/di', '../directives/message-event'], function(exports_1) {
+};System.register(['angular2/angular2', 'angular2/di', 'perfmjs/angular2/directives/message-event'], function(exports_1) {
     var angular2_1, di_1, message_event_1;
     var MyDirective, RedDec, SsqBobao;
     return {